api/requirements.txt
Django==4.2.16
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
django-cors-headers